Rename hardware to hardware_button
