+++ src/views/mod.rs
@@ -1,5 +1,5 @@
 pub mod configuration_column;
-pub mod hardware;
+pub mod hardware_button;
 pub mod info;
 pub mod layout_selector;
 pub mod main_row;
